+++ MyCarService/Interfaces/IVehicleRepository.cs
@@ -4,6 +4,6 @@ namespace MyCarService.Interfaces
 {
     public interface IVehicleRepository : IGenericRepository<Vehicle>
     {
-        IEnumerable<Vehicle> GetAllOwnerVehicles(int ownerId);
+        IEnumerable<Vehicle> GetAllOwnerVehicles(long ownerId);
     }
 }
