+++ MyCarService/Models/CarModel.cs
@@ -2,7 +2,7 @@
 {
     public class CarModel
     {
-        public int ModelId { get; set; }
+        public long ModelId { get; set; }
         public string? MakeName { get; set; }
 
         public string? ModelName { get; set; }
