+++ MyCarService/Models/DatabaseModels/Service.cs
@@ -2,8 +2,8 @@
 {
     public class Service
     {
-        public int Id { get; set; }
-        public int VehicleId { get; set; }
+        public long Id { get; set; }
+        public long VehicleId { get; set; }
         public int Millage { get; set; }
         public string? ServiceType { get; set; }
         public string? Description { get; set; }
