+++ MyCarService/Repositories/GenericRepository.cs
@@ -31,7 +31,7 @@ namespace MyCarService.Repositories
         {
             return _context.Set<T>().ToList();
         }
-        public T? GetById(int id)
+        public T? GetById(long id)
         {
             return _context.Set<T>().Find(id);
         }
